vista de errores

+++ Proyecto_Ciudad_De_Los_Ninos/Proyecto_Ciudad_De_Los_Ninos/Controllers/IncidentesController.cs
@@ -6,6 +6,7 @@ using API_Ciudad_De_Los_Ninos.Models;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Proyecto_Ciudad_De_Los_Ninos.Models;
 using Microsoft.AspNetCore.Authorization;
+using System.Diagnostics;
 
 namespace API_Ciudad_De_Los_Ninos.Controllers
 {
@@ -135,16 +136,37 @@ namespace API_Ciudad_De_Los_Ninos.Controllers
             return View(incidente);
         }
 
-        // POST: Incidentes/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var incidente = await _context.Incidentes.FindAsync(id);
-            _context.Incidentes.Remove(incidente);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            try
+            {
+                var incidente = await _context.Incidentes.FindAsync(id);
+                if (incidente == null)
+                {
+                    return RedirectToAction(nameof(Index), new { errorMessage = "El incidente no se encontró." });
+                }
+
+                _context.Incidentes.Remove(incidente);
+                await _context.SaveChangesAsync();
+
+                return RedirectToAction(nameof(Index));
+            }
+            catch (Exception)
+            {
+                var errorViewModel = new ErrorViewModel
+                {
+                    StatusCode = 500,
+                    Message = "Ocurrió un error al intentar eliminar el incidente.",
+                    RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier
+                };
+
+                return View("Error", errorViewModel);
+            }
         }
+
+
 
         private bool IncidenteExists(int id)
         {
